Split Mood enum to the own file
