+++ sinner/models/audio/PygameAudioBackend.py
@@ -5,7 +5,7 @@ from argparse import Namespace
 from moviepy.editor import AudioFileClip
 import pygame
 
-from sinner.models.logger.Status import Mood
+from sinner.models.logger.Logger import Mood
 from sinner.models.audio.BaseAudioBackend import BaseAudioBackend
 from sinner.utilities import get_file_name, normalize_path
